update reconfigure-pipeline installation to be idempotent

+++ packages/reconfigure-pipeline.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 . ~/workspace/log-cache-dotfiles/support/helpers.sh
 
@@ -7,17 +7,17 @@ VERSION=0.22.0
 
 echo_installing
 
-# reconfigure-pipelines doesn't currently output a version string anywhere.
-# we'll just reinstall it every time.
-#if [[ `reconfigure-pipeline --version` = "reconfigure-pipeline version $VERSION" ]]; then
-#    echo "Looks like reconfigure-pipeline $VERSION is already installed, skipping..."
-#    exit
-#fi
+if [ -f /usr/local/share/reconfigure-pipeline-$VERSION ]; then
+    echo "Looks like reconfigure-pipeline $VERSION is already installed, skipping..."
+    exit
+fi
 
+rm /usr/local/share/reconfigure-pipeline-*
 wget -O reconfigure-pipeline-$VERSION.tar.gz https://github.com/pivotal-cf/reconfigure-pipeline/releases/download/v$VERSION/reconfigure-pipeline-linux.tar.gz
 tar -xzvf reconfigure-pipeline-$VERSION.tar.gz
-sudo mv reconfigure-pipeline /usr/local/bin
-sudo chmod +x /usr/local/bin/reconfigure-pipeline
+sudo mv reconfigure-pipeline /usr/local/share/reconfigure-pipeline-$VERSION
+sudo chmod +x /usr/local/share/reconfigure-pipeline-$VERSION
+ln -s /usr/local/share/reconfigure-pipeline-$VERSION /usr/local/bin/reconfigure-pipeline
 rm reconfigure-pipeline-$VERSION.tar.gz
 
 echo_installed
